Extract Soil Moisture Data from TERN's SMIPS Using Australian GPS Coordinates

Extracts the soil moisture at the GPS points provided assuming that they are land-based coordinates in Australia from the Soil Moisture Integration and Prediction System SMIPS data set.

Usage

extract_smips(
  x,
  data = "smips",
  collection = "totalbucket",
  day,
  api_key = nert::get_key(),
  max_tries = 3,
  initial_delay = 1
)

Arguments

x

List. An object with names containing vector pairings of longitude and latitude values expressed as decimal degree values in that order, each individual vectors' items should be named "x" (longitude) and "y" (latitude), respectively. The list item names should be descriptive of the individual vectors and will be included in a "location" column of the output.

data

A character vector of the data source to be queried, currently only “smips”.

collection

A character vector of the data collection to be queried, currenly only “smips” is supported with the following collections:

SMindex

bucket1

bucket2

deepD

runoff

totalbucket

Defaults to “totalbucket”. Multiple collections are supported, e.g., collection = c("SMindex", "totalbucket").

day

A vector of date(s) to query, e.g., day = "2017-12-31" or day = seq.Date(as.Date("2017-12-01"), as.Date("2017-12-31"), "days"), both Character and Date classes are accepted.

api_key

A character string containing your API key, a random string provided to you by TERN, for the request. Defaults to automatically detecting your key from your local .Renviron, .Rprofile or similar. Alternatively, you may directly provide your key as a string here or use functionality like that from keyring. If nothing is provided, you will be prompted on how to set up your R session so that it is auto-detected and a browser window will open at the TERN website for you to request a key.

max_tries

An integer Integer with the number of times to retry a failed download before emitting an error message. Defaults to 3.

initial_delay

An Integer with the number of seconds to delay before retrying the download. This increases as the tries increment. Defaults to 1.

Value

A data.table::data.table with the provided GPS coordinates and the respective soil moisture data from SMIPS and coordinates from the SMIPS ("smips_longitude" and "smips_latitude")

References

https://portal.tern.org.au/metadata/TERN/15728dba-b49c-4da5-9073-13d8abe67d7chttps://portal.tern.org.au/metadata/TERN/d1995ee8-53f0-4a7d-91c2-ad5e4a23e5e0

Examples

if (FALSE) { # interactive()
locs <- list(
  "Merredin" = c(x = 118.28, y = -31.48),
  "Corrigin" = c(x = 117.87, y = -32.33),
  "Tamworth" = c(x = 150.84, y = -31.07)
)

extract_smips(x = locs, day = "2023-09-23")
}